Set up an empty Rust static library
We'd like to start using Rust in some of our binaries as a way to
improve robustness and memory safety.  To start, set up a workspace,
which contains only a single crate, git-internal, which builds a static
library.  Update the Makefile to build this library using Cargo and link
it into our binary.

For the moment, leave this library empty.  In future commits, we'll
introduce more functionality, but this serves as an easy test case to
demonstrate how the functionality works.

Finally, add functionality to specify the features used.  Right now,
we're doing nothing useful, so we are not setting any features.
However, in the future, this will allow us to use conditional
compilation to either include or exclude functionality, such as gettext.