rust: add support for localization
Git has been translated into many languages.  As we start using Rust,
we'll want to avoid regressing this functionality.  Add support for
translation in Rust by using gettext, if NO_GETTEXT is not defined in
the Makefile, or a stub implementation if it is.

Add a basic unit test that we get something as a response, which make
sure whatever implementation we're using is basically functional.  Use
an existing string to avoid burdening translators.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>