rust: provide conversion between byte strings and OS strings
Git serializes all of its paths and other strings as byte strings,
possibly in UTF-8.  However, Windows natively uses potentially
ill-formed UTF-16 for its strings (an encoding known as WTF-16).  Rust
exposes the OS-specific strings as the struct OsStr.

However, it does not provide a way to consistently convert back and
forth between byte strings and OS-specific strings, which we will
obviously need to do.  In 1.74, which is too new for us, Rust does
provide a way to encode these strings, but it's not guaranteed to be
consistent between Rust versions, and we absolutely want to process
strings in a consistent way regardless of the version in use.

To do that, add some traits that can convert back and forth between byte
strings and OsStr and friends.  In addition, in the future we'll want to
be able to format error messages which may contain a byte string, so
check if the string is valid UTF-8, and if not, print it in an escaped
form.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>