rust: enable exporting functions to C
Rust makes it possible to define functions using the C API, which is
convenient.  However, by default, one has to define the C functions and
types by hand, which is decidedly less convenient.

Fortunately, there is a Rust crate called cbindgen that crawls Rust code
and looks for definitions that can safely be exported via the C ABI, and
it emits a header for them.  The actual code is written in Rust and
provided by the static library we link into the binaries.

Provide a configuration file for cbindgen that tells it we want C, not
C++, definitions and that we want git-compat-util.h, and not the default
system headers, to be included.  Also add an include guard to avoid any
problems, and ignore the file, since it is generated.

Note that because this file is generated at build time, we need a
build.rs file, which performs that task and will be automatically
invoked by Cargo when building.

At the moment, we don't have any definitions to export, but that will
change in a future commit.

Note that any syntax error will fail to build the bindings and therefore
fail the build process.  This is expected, but it prevents us from
getting a useful error message about our syntax error from the compiler
since it isn't invoked.  To help work around that, add an environment
variable, SKIP_BINDINGS, that can be used in case of a syntax error to
get better diagnostics.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>