rust: add an error type
Errors are very common in a wide variety of code, and Git is no
exception.  We have very little code that can currently fail, but our
encoding and decoding routines might encounter an invalid UTF-8
character on Windows, which we'd want to handle.

Add an error type that we can use throughout the code, and add an error
code for the invalid UTF-8 case that we have.  Allow converting it
automatically.  In addition, add a variant for annotating other error
codes in a useful way.  Also provide an enum that will be automatically
emitted in C so that we can check error codes there as well.  Rename the
error type to "Err", which is short and also provides us with C values
like ERR_INVALID_UTF8, which is exactly the form we want.

Note that the use of __ is handled by xgettext just as the function
gettext is, and it's shorter.  Since we have the possibility of having
many errors here, use the shorter variant.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>