rust: define a basic config parser
We already have a configuration parser in C.  However, it suffers from a
few flaws which we can improve on.

First, C is not a memory-safe language and our config parser is
frequently used on untrusted data (such as .gitmodules files).  If we're
not very careful, we could accidentally introduce a security-sensitive
bug which might lead to remote code execution.  We already know that it
doesn't handle very large config files, for instance, which, while
uncommon, should not intrinsically be a problem.

Second, we don't have a lot of tests for invalid configuration files.
We can improve on that and verify a large collection of error cases (in
parallel, no less) by writing tests.

Third, the style of our configuration parser callback lends itself to
storing the configuration in global variables, which we do all over the
code.  This is not very tidy, and it makes libification difficult.
Improve on this design by allowing callbacks to return the value they've
parsed and retain them for future use.  This allows us to incrementally
change the callbacks into Rust and then preserve the parsed data in a
map, avoiding the need for global variables.

Move several of the enums and structs which we use in C into Rust and
export their types.  This allows the C code to keep working while we
develop the Rust parser in parallel.  Note that the Rust parser does not
yet parse and this is only a scaffold, but a future commit will
introduce actual parsing functionality.

Use a BufReader when reading from our arbitrary read source to avoid
calling read(2) on each byte, which would be wildly inefficient.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>