rust: add a trait to make items localizable
Rust provides native support for the operating system's error values via
the std::io::Error type.  However, Rust doesn't provide localization and
we want our errors to be appropriately localized.

Add a trait that can extract the raw OS error from a type and pass it
through strerror so that we get a localized value.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>