rust: add error types for parsing config
In a future commit, we'll add support for parsing config files in Rust
instead of in C and we'll want to be able to handle multiple different
types of errors.  Add three new error types that we can use to handle
various parsing problems.

For the UnexpectedConfigData message, keep the error messages similar to
those to those used in C, but provide a more descriptive reason as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>