WIP setup: allow specifying the compat hash algo by environment
We'd like to be able to enable creating repositories with both a regular
and a compatibility hash algorithm.  Allow the user to specify
GIT_DEFAULT_HASH with two algorithms separated by a colon, where the
regular is the former and the compatibility is the latter.

WIP: should not be enabled until ready for production

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>