WIP clone: enable hash algorithm interop on clone
When we set up a new repository by cloning and we have enable algorithm
interoperability, enable the interoperability mode on the new repository
as well.  This allows both to function in a reasonable way and it
additionally makes several of our tests pass where they otherwise would
not.

WIP: the hard-coded 3 is hideous
WIP: this should be configurable both via environment and config
WIP: this should be partially controlled by whether the remote algorithm
     and our current algorithm are the same once we allow people to
     choose

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>