t7004: fix signature parsing in hash compat mode
When we're running the tests with a compatibility hash, our tags are
double signed automatically.  That means that there will be a signature
for both the SHA-1 format and the SHA-256 format, so either strip this
extra signature or don't add a new one when running the tests.  However,
do verify in this mode that the tag is in fact double signed.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>