loose: don't try to insert objects in invalid cache
Our loose object cache is typically created when we load the loose
object map, but we can clear it if someone calls reprepare_packed_git.
In such a case, the data will be freed and the pointer will be reset to
NULL.  If we then attempt to insert an object into the loose object map,
we'll end up crashing due to a NULL pointer dereference.

The easiest way around this is merely to check the pointer before
inserting, so do that.  This fixes git am in hash compatibility mode.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>