loose: ensure loose object map is initialized
In certain situations, we can end up with an entry in the object
database list where the loose object map is not initialized.  If we
then attempt to add an object, we segfault due to attempting to
dereference a NULL pointer.

To avoid this, ensure that the loose object map is initialized.  This
avoids segfaults in git add when in hash compatibility mode.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>