index-pack: allow specifying compat object format
Right now, if we're in hash compatibility mode, we'll generate an index
with both hash algorithms in it.  However, when we're outside of a
repository, there's no way to indicate that we want both algorithms,
which makes verification fail due to the verified data containing only
the primary algorithm and therefore mismatching.

To make this work in a graceful way, add a --compat-object-format option
and set the compatibility object format just like we do for the main
one.  Update t1050 to test this case, which also make the test pass in
hash compatibility mode.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>