serve: allow serving multiple instances of a capability
Right now, in protocol v2, it's only possible to serve one instance of a
capability.  However, in the future, we'll allow the server to possibly
serve multiple object-format capabilities if it has interoperability
enabled and can send data in either format.

To make this possible, add an additional integer argument that starts at
zero and increases each call.  Once there are no more instances of the
capability to be handled, the advertise function can return 0 to
indicate this fact.  Update all of the existing advertise functions to
immediately return 0 if the integer argument is nonzero, since we do not
yet offer any capabilities that need multiple advertisements.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>