WIP: serve: advertise compatibility hash algorithm
Right now, we only ever advertise the main hash algorithm in the
capabilities advertisement.  However, now that we support
hash compatibility, we can advertise both algorithms and let the client
choose which algorithm they'd like to receive the ref advertisement and
the data in.

We always advertise the main hash algorithm first, since older clients
will always pick the first algorithm.  It's also important to note that
we send two object-format lines instead of a single
"object-format=sha256 sha1" line, since older implementations will
attempt to parse that as a single hash algorithm and choke.  The design
used here provides good backwards compatibility with older clients.

We also offer a config option, transfer.advertisecompatobjectformat,
that defaults to true but can be disabled.  This is because some servers
will want to carry both object formats but not go through the expense of
converting data on the server side, leaving the client to perform that
operation.

WIP: squash into following commit

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>