ls-refs: allow converting ref advertisements to compat hash
Right now, we always advertise refs in the main hash algorithm.
However, if we have the compatibility mapping, we can advertise them in
the compatibility hash instead if the client requests it.

Store the hash that the client requested in the ls-refs data and use it
to map the object ID.  Default to the hash algorithm from the packet
reader, which in turn defaults to the main hash algorithm.  Note that
repo_oid_to_algop returns early without doing anything expensive if the
hash algorithm is already correct, so this is functionally a no-op if
the main hash algorithm (or no hash algorithm) is selected.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>