connect: allow search for multiple features with same name
In a future commit, we'll want to read multiple object-format headers
with server_supports_feature.  Right now, it exits as soon as it finds
the first entry with that feature, but in the future, we'll want to find
the second or subsequent instance as well.  Don't skip the remainder of
the features immediately upon finding the first instance of the feature.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>