pack-objects: add a variable for the hash algorithm
We currently pack all objects in the main hash algorithm of the
repository.  However, in the future, we'd like to be able to pack
objects in the compatibility hash algorithm as well.  To make this
easier, add a variable that represents the hash algorithm we're using
in this pack, default it to the main hash algorithm, and use it when
hashing the pack contents.

In a future commit, we'll add support for specifying a different hash
algorithm as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>